Add Log4j implementation of SLF4J to the webapp
This was previously included indirectly via Jena.

Also, remove the Log4j stuff for jetty:run. Not clear what the
original intent was, and now it seems to have no effect. Jetty is
logging to the Maven logging system, and it doesn't look like that
can be changed by adding jars to the plugin classpath.